+++ okai/packages/client-auto/src/index.ts
@@ -1,4 +1,4 @@
-import { Client, IAgentRuntime } from "@okcashpro/okai";
+import { Client, IAgentRuntime, okaiLogger } from "@okcashpro/okai";
 
 export class AutoClient {
     interval: NodeJS.Timeout;
@@ -10,7 +10,7 @@ export class AutoClient {
         // start a loop that runs every x seconds
         this.interval = setInterval(
             async () => {
-                console.log("running auto client...");
+                okaiLogger.log("running auto client...");
             },
             60 * 60 * 1000
         ); // 1 hour in milliseconds