+++ okai/packages/plugin-evm/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.7-alpha.1](https://github.com/okcashpro/okai/compare/v0.1.7-alpha.0...v0.1.7-alpha.1) (2024-12-18)
+
+**Note:** Version bump only for package @okcashpro/plugin-evm
+
+
+
+
+
+## [0.1.7-alpha.0](https://github.com/okcashpro/okai/compare/v0.1.5-alpha.10...v0.1.7-alpha.0) (2024-12-18)
+
+
+### Bug Fixes
+
+* fix rpcError for ethereum mainnet ([5e7efb0](https://github.com/okcashpro/okai/commit/5e7efb0d9cbc31a1aa8526730362ff4020be05b2))
+* fix undefined chain ([ee0a3aa](https://github.com/okcashpro/okai/commit/ee0a3aadc359c87bede143c08506abac51e809b0))
+* improve chain selection ([9a33703](https://github.com/okcashpro/okai/commit/9a33703e61d84bca962830cab15b442f55e680a3))
+* **plugin-evm:** eth transfer action ([9aa80fa](https://github.com/okcashpro/okai/commit/9aa80fa37f96baca88f762d5b3b1b5f03d090e36))
+* remove duplicated import ([8c15447](https://github.com/okcashpro/okai/commit/8c15447d88ac37d1a9bc15d1053250bf797dd761))
+* throw if evm private key is missing ([9fc4b0e](https://github.com/okcashpro/okai/commit/9fc4b0ea7df91541b6c5853daf9bfe9c73a170db))
+* transfer action fix [#735](https://github.com/okcashpro/okai/issues/735) ([b95bf03](https://github.com/okcashpro/okai/commit/b95bf03d3b6b2ace41c0975a21d8d086c1ce45af))
+* transfer action fix for EVM Plugin can't run any action [#735](https://github.com/okcashpro/okai/issues/735) ([ee43b45](https://github.com/okcashpro/okai/commit/ee43b4560fb907298173b096c8a815874a5f6d51))
+
+
+### Features
+
+* add better feedback from wallet provider ([a420a96](https://github.com/okcashpro/okai/commit/a420a966ae3b9b9b064ee525164b5379dc1b8bc3))
+* add custom rpc url support ([c6a3531](https://github.com/okcashpro/okai/commit/c6a3531fc8cc918214f15985c5223ea1bc138179))
+
+
+
+
+
 ## [0.1.5-alpha.10](https://github.com/okcashpro/okai/compare/v0.1.5-alpha.9...v0.1.5-alpha.10) (2024-12-11)
 
 **Note:** Version bump only for package @okcashpro/plugin-evm