+++ okai/packages/plugin-node/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.7-alpha.1](https://github.com/okcashpro/okai/compare/v0.1.7-alpha.0...v0.1.7-alpha.1) (2024-12-18)
+
+**Note:** Version bump only for package @okcashpro/plugin-node
+
+
+
+
+
+## [0.1.7-alpha.0](https://github.com/okcashpro/okai/compare/v0.1.5-alpha.10...v0.1.7-alpha.0) (2024-12-18)
+
+
+### Bug Fixes
+
+* load image from diff endpoints ([4ce5f59](https://github.com/okcashpro/okai/commit/4ce5f59a0cac3acf31b914194960b476e5d308bc))
+
+
+### Features
+
+* Add plugin-nft-generation: support for creating Solana NFT collections. ([3ff0677](https://github.com/okcashpro/okai/commit/3ff0677f3534f11bdc9e29843f1c936662e3d299))
+
+
+
+
+
 ## [0.1.5-alpha.10](https://github.com/okcashpro/okai/compare/v0.1.5-alpha.9...v0.1.5-alpha.10) (2024-12-11)
 
 **Note:** Version bump only for package @okcashpro/plugin-node